python/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
beautifulsoup4==4.12.2   
boto3==1.34.5
botocore==1.34.5
Brotli==1.1.0
certifi==2023.11.17      
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
databases==0.8.0
easydict==1.11
exceptiongroup==1.2.0    
fastapi==0.105.0
filelock==3.13.1
fsspec==2023.12.2        
gdown==4.7.1
greenlet==3.0.2
h11==0.14.0
huggingface-hub==0.20.1
idna==3.6
Jinja2==3.1.2
jmespath==1.0.1
kornia==0.7.0
MarkupSafe==2.1.3
mpmath==1.3.0
mutagen==1.47.0
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
pycryptodomex==3.19.0
pydantic==2.5.2
pydantic_core==2.14.5
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.9.0
safetensors==0.4.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.50
starlette==0.27.0
sympy==1.12
timm==0.9.12
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
transparent-background==1.2.9
typing_extensions==4.9.0
urllib3==2.0.7
uvicorn==0.24.0.post1
websockets==12.0
yt-dlp==2023.11.16